requirements.txt
audiotsm==0.1.2
MouseInfo==0.1.3
numpy==1.21.3
opencv-python==4.7.0.68
Pillow==8.4.0
PyAutoGUI==0.9.53
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
pytube==11.0.1
pytweening==1.0.4
scipy==1.7.1
